Move runSearch to StatsUtils.js and refactor scripts using it
